# static/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Wallet Organization
  version: 0.0.1
  description: |
    Wallet Organization Service is a REST HTTP API for managing wallet organizations. A wallet organization is a company
    or entity that is registered into the wallet platform ant that uses the wallet platform features.
    Wallet Organization Service exposes REST endpoint for:

    * Creating wallet organizations
    * Retrieving wallet organization Information
    * Updating wallet organizations

    # Security

    This service is not publicly accessible. Only the Wallet Gateway Service can make calls to its endpoints

    ## Authentication

    Access to the API is controlled by an OpenID token
    The client needs to request an access token from the oauth2 server. It provides its client id and client secret.
    Then all requests to the Wallet Account User Service must be made over HTTP/HTTPS with this access token

    # Responses

    ## HTTP Response Codes
    | Status Code | Usage                                                       |
    |-------------|-------------------------------------------------------------|
    | 200         | The request completed successfully                          |
    | 201         | A new resource has been created                             |
    | 400         | Malformed request                                           |
    | 401         | Missing access token                                        |
    | 403         | The client does not have permission to access this resource |
    | 404         | Resource not found                                          |
    | 405         | Method not supported                                        |
    | 500         | Internal server error                                       |
    | 503         | Service unvailable                                          |

    ## Errors
    whenever an error occured within the platform, an error response that describes the problem will be returned

    | Field Name        | Field Description     |
    |-------------------|-----------------------|
    | Status            | "fail" or "error"     |
    | StatusCode        | A business error code |
    | StatusDescription | Error description     |

    ## Error codes

    | Error Code | Error Description                                |
    |------------|--------------------------------------------------|
    | 6001       | Can't obtain list of wallet organizations        |
    | 6010       | Can't create wallet organization                 |
    | 6020       | Wallet organization not found                    |
    | 6030       | Can't obtain wallet organization data            |
    | 6040       | Unable to update wallet organization data        |
    | 6000       | Invalid Request                                  |
    | 6050       | System Error                                     |

    # Changelog

    ## 2019-01-05
    * Created .....

    # Useful links
  termsOfService: 'http://swagger.io/terms/'
  x-logo:
    url: '/static/img/NBK_logo.jpeg'
    backgroundColor: '#FFFFFF'
    altText: NBK Logo
servers:
  - url: 'https://wallet-account-organization-svc-fjhmnd5asa-ew.a.run.app'
    description: Testing Server
tags:
  - name: Wallet Organizations
    description: Wallet Organizations are used to handle operations on wallet organizations
    externalDocs:
      url: 'https://wallet-gateway-svc-uzqgb7lfiq-ew.a.run.app'
paths:
  /v1/organizations:
    get:
      summary: Obtain a list of wallet organizations
      tags:
        - Wallet Organizations
      responses:
        '200':
          description: Successfully obtained list of wallet organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/ResponseStatus'
                  data:
                    type: object
                    properties:
                      walletOrganizations:
                        $ref: '#/components/schemas/WalletOrganizationsWithId'

        '500':
          description: Error when obtaining list of wallet organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: fail
                statusCode: 6001
                statusDescription: Error when obtaining list of wallet organizations

    post:
      summary: Create a Wallet Organization
      tags:
        - Wallet Organizations
      requestBody:
        description: The wallet Organization specification
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletOrganization'
      responses:
        '201':
          description: Successfully created wallet organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/ResponseStatus'
                  data:
                    type: object
                    properties:
                      walletOrganizationWithId:
                        $ref: '#/components/schemas/WalletOrganizationWithId'
        '500':
          description: Error when creating wallet organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: fail
                statusCode: 6010
                statusDescription: Error when creating wallet organization
  /v1/organizations/{organizationId}:
    get:
      summary: Fetch wallet organization data
      tags:
       - Wallet Organizations
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            example: 6784bbea-ea94-426e-bb64-594eeee6cb03
      responses:
        '200':
          description: Successfully fetched wallet organization data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/ResponseStatus'
                  data:
                    type: object
                    properties:
                      walletOrganization:
                        $ref: '#/components/schemas/WalletOrganizationWithId'
        '404':
          description: wallet organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: fail
                statusCode: 6020
                statusDescription: wallet organization not found
        '500':
          description: Error when fetching wallet organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: fail
                statusCode: 6021
                statusDescription: Error when fetching wallet organization
    patch:
      summary: Update wallet organization data
      tags:
        - Wallet Organizations
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            example: 6784bbea-ea94-426e-bb64-594eeee6cb03
      requestBody:
        description: The wallet organization specification
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletOrganization'
      responses:
        '201':
          description: Successfully updated wallet organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/ResponseStatus'
                  data:
                    type: object
                    properties:
                      walletOrganization:
                        $ref: '#/components/schemas/WalletOrganizationWithId'
        '500':
          description: Error when updating the wallet account organization information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: fail
                statusCode: 4030
                statusDescription: Error when updating the wallet account organization information

components:
  schemas:
    ResponseStatus:
      type: string
      enum:
        - success
        - fail
        - error
    ErrorResponseStatus:
      type: string
      enum:
        - fail
        - error
    ResponseStatusCode:
      type: number
      example: 1023
    ResponseStatusDescription:
      type: string
      example: The wallet plan does not exist within the system
    ErrorResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ErrorResponseStatus'
        statusCode:
          $ref: '#/components/schemas/ResponseStatusCode'
        statusDescription:
          $ref: '#/components/schemas/ResponseStatusDescription'
    Address:
      type: object
      properties:
        streetName:
          type: string
          example: Astra House
        streetNumber:
          type: string
          example: 201
        city:
          type: string
          example: New Cross
        postCode:
          type: string
          example: SE14 24
        state:
          type: string
          example: London
        country:
          type: string
          example: United Kingdom
    WalletOrganization:
      type: object
      properties:
        name:
          type: string
          description: wallet organization name
          example: NBK
        address:
          $ref: '#/components/schemas/Address'
        email:
          type: string
          description: wallet organization email address
          format: email
          example: my-email@gmail.com
        phoneNumber:
          type: string
          description: wallet organization phone number
          example: 00447307568957
        mobileNumber:
          type: string
          description: wallet organization mobile number
          example: 00447307568957
        applications:
          $ref: '#/components/schemas/WalletClientApplications'
        walletOrganizationDocuments:
          $ref: '#/components/schemas/WalletOrganizationDocuments'
    WalletOrganizationWithId:
      allOf:
        - $ref: '#/components/schemas/WalletOrganization'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              description: wallet organization identifier
              format: uuid
              example: 45466ea6-512d-42c7-b57f-d11c53890456
    WalletOrganizations:
      type: array
      items:
        $ref: '#/components/schemas/WalletOrganization'
    WalletOrganizationsWithId:
      type: array
      items:
        $ref: '#/components/schemas/WalletOrganizationWithId'
    WalletClientApplication:
      type: object
      properties:
        name:
          type: string
          description: Application name
          example: NBK App
        credentials:
          type: object
          properties:
            oauth2:
              type: object
              properties:
                clientId:
                  type: string
                  description: OAuth2 client identifier
                  example: 1t37i9t15h3rvlib7g1u7odp23
                clientSecret:
                  type: string
                  description: oauth2 client secret
                  example: avqbjl9vfeo1spfhv5qfp4ojrplg6guf3gv44q1hpvffk6nab8g
    WalletClientApplications:
      type: array
      items:
        $ref: '#/components/schemas/WalletClientApplication'
    WalletOrganizationDocument:
      type: object
      properties:
        documentId:
          type: string
          description: Wallet document identifier
          format: uuid
          example: 45466ea6-512d-42c7-b57f-d11c53890456
    WalletOrganizationDocuments:
      type: array
      items:
        $ref: '#/components/schemas/WalletOrganizationDocument'